requirements.txt
opencc-python-reimplemented==0.1.7
pandas==2.0.3
openpyxl==3.1.3
requests==2.31.0
tqdm==4.66.4
langchain==0.2.3
langchain-openai==0.1.8
openai==1.33.0
rouge-chinese==1.0.3
jieba==0.42.1
sentence-transformers==3.0.1